Google Natural Sentiment thing
Note: You'll have to redo credentials and stuff for the environment variable. I can help with that
Used eclipse for this. You'll also have to download Google Cloud libraries and include that in your package
Can help with this if you need. Just lmk (and please have eclipse downloaded). 

This page can help, if you have problems/trouble understanding hit it up. Find me if you need help, but I didn't look through this website that much (only used it for implementing and making the program work).
https://cloud.google.com/natural-language/docs/quickstart